backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
fredapi
yfinance
numpy
pandas
cvxpy
optuna
python-dotenv
